Split the CD ripper into two classes
Separate the cd ripper dialog from the ripper itself.
