+++ src/ripper/ripcddialog.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>RipCD</class>
- <widget class="QDialog" name="RipCD">
+ <class>RipCDDialog</class>
+ <widget class="QDialog" name="RipCDDialog">
   <property name="windowModality">
    <enum>Qt::NonModal</enum>
   </property>
